Rewrite dynamic Groovy call interception to make it more generic
Groovy has a variety of ways to call methods/constructors or read
properties. One has to override a lot of methods of the CallSite to
properly intercept them: call() to intercept instance and static
methods, callStatic() to intercept calls to statically imported methods,
callGetProperty() to intercept property reads. There are "optimized"
calls to methods with separate parameters but there are also bulk
overloads that accept argument arrays. The latter are used for spread
calls (e.g. `foo(*[bar, baz])`).

This diversity caused our instrumentation to miss some cases, most
notably spread calls were not intercepted at all, and some calls weren't
interceped if static imports are used.

This commit separates the generic "interception" code that deals with
Groovy's CallSites and their arguments from the user code that performs
actions when the intercepted methods are called. The interceptor takes
care of overriding necessary methods, unwrapping the arguments, and
provides a way to call the original Groovy implementation if the user
code decides that the invocation shouldn't be intercepted.

The Invocation object hides the details of the actual Groovy machinery
involved in calling the method/constructor, so the interceptor can be
extended further to support Groovy's invokedynamic-based implementation
while reusing the user code as is.

One downside compared to the previous implementation is that optimized
code paths with separate arguments are no longer optimized as the
arguments are always packed into arrays before further processing.
It shouldn't be a big problem in practice as the "interesting" calls are
not that frequent and others are inlined by the Groovy runtime after the
first invocation. A proper microbenchmarking coverage is necessary to
optimize things further.